requirements.txt
asgiref==3.8.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.8
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.11
pillow==12.0.0
pycparser==2.23
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.5.0
